# be/docs/swagger.yaml
basePath: /api
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorData'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.ErrorData:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  models.FlagRequest:
    properties:
      is_flagged:
        type: boolean
      notes:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@form.com
    name: API Support
  description: API for form application with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: form API
  version: "1.0"
paths:
...
  /forms/{form_id}/responses:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all responses for a form with filtering
        and sorting options
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: submitted_at
        description: Sort by field
        enum:
        - submitted_at
        - score
        - time_spent
        - respondent_name
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: Search by respondent name or email
        in: query
        name: search
        type: string
      - description: Filter by pass/fail status
        in: query
        name: is_passed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all responses for a form
      tags:
      - form-responses
  /forms/{form_id}/responses/{response_id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a form response and all associated answers
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Response ID
        in: path
        name: response_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a specific response
      tags:
      - form-responses
    get:
      consumes:
      - application/json
      description: Retrieve complete details including all answers for a specific
        form response
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Response ID
        in: path
        name: response_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get detailed information for a specific response
      tags:
      - form-responses
  /forms/{form_id}/responses/{response_id}/flag:
    put:
      consumes:
      - application/json
      description: Mark a response as flagged for review or remove flag
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Response ID
        in: path
        name: response_id
        required: true
        type: string
      - description: Flag request with is_flagged and optional notes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FlagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Flag or unflag a response
      tags:
      - form-responses
  /forms/{form_id}/responses/summary:
    get:
      consumes:
      - application/json
      description: Retrieve summary statistics including total responses, average
        score, and pass rate
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get response summary statistics for a form
      tags:
      - form-responses
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"